+++ scripts/git/diff.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-git diff --stat --name-only HEAD^ HEAD